Move remaining global reconstruction code
